商品削除機能提出2回目

+++ app/controllers/items_controller.rb
@@ -37,9 +37,13 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if current_user == @item.user
+      item = Item.find(params[:id])
+      item.destroy
+      redirect_to root_path
+    else
+      render :show
+    end
   end
 
   private
